@@ -4,7 +4,7 @@
 // By: SharkPool
 // License: MIT
 
-// Version V.1.0.07
+// Version V.1.0.09
 
 (function (Scratch) {
   "use strict";
@@ -53,6 +53,10 @@
   }
 
   // camera utils
+  const radianConstant = Math.PI / 180;
+  const epsilon = 1e-12;
+  const applyEpsilon = (value) => (Math.abs(value) < epsilon ? 0 : value);
+
   function setupState(drawable) {
     drawable[cameraSymbol] = {
       name: "default",
@@ -65,23 +69,26 @@
 
   function translatePosition(xy, invert, camData) {
     if (invert) {
-      const invRads = (camData.ogDir / 180) * Math.PI;
+      const invRads = camData.ogDir * radianConstant;
       const invSin = Math.sin(invRads),
         invCos = Math.cos(invRads);
       const scaledX = xy[0] / camData.ogSZ;
       const scaledY = xy[1] / camData.ogSZ;
       const invOffX = scaledX * invCos + scaledY * invSin;
       const invOffY = -scaledX * invSin + scaledY * invCos;
-      return [invOffX - camData.ogXY[0], invOffY - camData.ogXY[1]];
+      return [
+        applyEpsilon(invOffX - camData.ogXY[0]),
+        applyEpsilon(invOffY - camData.ogXY[1]),
+      ];
     } else {
-      const rads = (camData.dir / 180) * Math.PI;
+      const rads = camData.dir * radianConstant;
       const sin = Math.sin(rads),
         cos = Math.cos(rads);
       const offX = xy[0] + camData.xy[0];
       const offY = xy[1] + camData.xy[1];
       return [
-        camData.zoom * (offX * cos - offY * sin),
-        camData.zoom * (offX * sin + offY * cos),
+        applyEpsilon(camData.zoom * (offX * cos - offY * sin)),
+        applyEpsilon(camData.zoom * (offX * sin + offY * cos)),
       ];
     }
   }
@@ -113,9 +120,10 @@
   }
 
   function updateCamera(camera) {
-    for (let i = 0; i < render._allDrawables.length; i++) {
-      const drawable = render._allDrawables[i];
-      if (!drawable || !drawable.getVisible() || !drawable.skin) continue;
+    for (let i = 0; i < render._drawList.length; i++) {
+      const drawableId = render._drawList[i];
+      const drawable = render._allDrawables[drawableId];
+      if (!drawable.getVisible() || !drawable.skin) continue;
       if (!drawable[cameraSymbol]) setupState(drawable);
 
       const camSystem = drawable[cameraSymbol];
@@ -192,11 +200,15 @@
     if (!this[cameraSymbol]) setupState(this);
     const camSystem = this[cameraSymbol];
     const thisCam = allCameras[camSystem.name];
+    let shouldEmit = false;
     if (camSystem.needsRefresh) {
       // invert camera transformations
       position = translatePosition(position, true, camSystem);
     }
 
+    shouldEmit =
+      camSystem.ogXY[0] !== thisCam.xy[0] ||
+      camSystem.ogXY[1] !== thisCam.xy[1];
     camSystem.ogXY = [...thisCam.xy];
     position = translatePosition(position, false, thisCam);
     if (camSystem.needsRefresh) {
@@ -207,7 +219,10 @@
         this._position[0] = position[0];
         this._position[1] = position[1];
       }
-      this.setTransformDirty();
+      if (shouldEmit) {
+        render.dirty = true;
+        this.setTransformDirty();
+      }
     } else {
       ogUpdatePosition.call(this, position);
     }
@@ -233,11 +248,13 @@
     if (!this[cameraSymbol]) setupState(this);
     const camSystem = this[cameraSymbol];
     const thisCam = allCameras[camSystem.name];
+    let shouldEmit = false;
     if (camSystem.needsRefresh) {
       // invert camera transformations
       const safeOgSZ = camSystem.ogSZ !== 0 ? camSystem.ogSZ : 1e-10;
       scale[0] /= safeOgSZ;
       scale[1] /= safeOgSZ;
+      shouldEmit = safeOgSZ !== thisCam.zoom;
     }
 
     // avoid dividing 0 by 0
@@ -249,7 +266,26 @@
     if (scale[1] === 0) scale[1] = 1e-10 * Math.sign(safeZoom);
 
     ogUpdateScale.call(this, scale);
-    this.skin?.emitWasAltered();
+    if (shouldEmit) {
+      this._renderer.dirty = true;
+      this._rotationCenterDirty = true;
+      this._skinScaleDirty = true;
+      this.setTransformDirty();
+    }
+  };
+
+  // Clones should inherit the parents camera
+  const ogInitDrawable = vm.exports.RenderedTarget.prototype.initDrawable;
+  vm.exports.RenderedTarget.prototype.initDrawable = function (layerGroup) {
+    ogInitDrawable.call(this, layerGroup);
+    if (this.isOriginal) return;
+
+    const parentSprite = this.sprite.clones[0]; // clone[0] is always the original
+    const parentDrawable = render._allDrawables[parentSprite.drawableID];
+    const name = parentDrawable[cameraSymbol]?.name ?? "default";
+
+    const drawable = render._allDrawables[this.drawableID];
+    bindDrawable(drawable, name);
   };
 
   // Turbowarp Extension Storage
@@ -604,13 +640,15 @@
         });
 
       // Custom Drawable Layer (CST's 3D or Simple3D Exts for Example)
-      for (var i = 0; i < render._allDrawables.length; i++) {
-        const drawable = render._allDrawables[i];
-        if (drawable !== undefined && drawable.customDrawableName !== undefined)
+      for (var i = 0; i < render._drawList.length; i++) {
+        const drawableId = render._drawList[i];
+        const drawable = render._allDrawables[drawableId];
+        if (drawable.customDrawableName !== undefined) {
           objectNames.push({
             text: drawable.customDrawableName,
-            value: `${i}=SP-custLayer`,
+            value: `${drawableId}=SP-custLayer`,
           });
+        }
       }
 
       // Sprites
@@ -700,10 +738,10 @@
     }
 
     translateAngledMovement(xy, steps, direction) {
-      const radians = direction * (Math.PI / 180);
+      const radians = direction * radianConstant;
       return [
-        xy[0] + steps * Math.cos(radians),
-        xy[1] + steps * Math.sin(radians),
+        applyEpsilon(xy[0] + steps * Math.cos(radians)),
+        applyEpsilon(xy[1] + steps * Math.sin(radians)),
       ];
     }
 
