@@ -1,13 +1,10 @@
-// Name: Cloudlink
+// Name: CloudLink V4
 // ID: cloudlink
 // Description: A powerful WebSocket extension for Scratch.
 // By: MikeDEV
 // License: MIT
 
-/* eslint-disable */
-// prettier-ignore
 (function (Scratch) {
-
   /*
   CloudLink Extension for TurboWarp v0.1.2.
 
@@ -37,9 +34,9 @@
   */
 
   // Require extension to be unsandboxed.
-  'use strict';
+  "use strict";
   if (!Scratch.extensions.unsandboxed) {
-    throw new Error('The CloudLink extension must run unsandboxed.');
+    throw new Error("The CloudLink extension must run unsandboxed.");
   }
 
   // Declare icons as static SVG URIs
@@ -81,7 +78,6 @@
 
   // Store extension state
   var clVars = {
-
     // Editor-specific variable for hiding old, legacy-support blocks.
     hideCLDeprecatedBlocks: true,
 
@@ -235,14 +231,24 @@
     handshakeAttempted: false,
 
     // Storage for the publically available CloudLink instances.
-    serverList: {},
-  }
+    serverList: {
+      0: {
+        id: "Localhost",
+        url: "ws://127.0.0.1:3000/",
+      },
+      7: {
+        id: "MikeDEV's Spare CL 0.2.0 Server",
+        url: "wss://cl.mikedev101.cc/",
+      },
+    },
+  };
 
   function generateVersionString() {
     return `${version.editorType} ${version.versionString}`;
   }
 
   // Makes values safe for Scratch to represent.
+  // eslint-disable-next-line require-await
   async function makeValueScratchSafe(data) {
     if (typeof data == "object") {
       try {
@@ -328,12 +334,14 @@
   function sendMessage(message) {
     // Prevent running this while disconnected
     if (clVars.socket == null) {
-      console.warn("[CloudLink] Ignoring attempt to send a packet while disconnected.");
+      console.warn(
+        "[CloudLink] Ignoring attempt to send a packet while disconnected."
+      );
       return;
     }
 
     // See if the outgoing val argument can be converted into JSON
-    if (message.hasOwnProperty("val")) {
+    if (Object.prototype.hasOwnProperty.call(message, "val")) {
       try {
         message.val = JSON.parse(message.val);
       } catch {}
@@ -341,28 +349,33 @@
 
     // Attach listeners
     if (clVars.listeners.enablerState) {
-
       // 0.1.8.x was the first server version to support listeners.
       if (clVars.linkState.identifiedProtocol >= 2) {
         message.listener = clVars.listeners.enablerValue;
 
         // Create listener
-        clVars.listeners.varStates[String(args.ID)] = {
+        clVars.listeners.varStates[message.listener] = {
           hasNew: false,
           varState: {},
           eventHatTick: false,
         };
-
       } else {
-        console.warn("[CloudLink] Server is too old! Must be at least 0.1.8.x to support listeners.");
+        console.warn(
+          "[CloudLink] Server is too old! Must be at least 0.1.8.x to support listeners."
+        );
       }
       clVars.listeners.enablerState = false;
     }
 
     // Check if server supports rooms
-    if (((message.cmd == "link") || (message.cmd == "unlink")) && (clVars.linkState.identifiedProtocol < 2)) {
+    if (
+      (message.cmd == "link" || message.cmd == "unlink") &&
+      clVars.linkState.identifiedProtocol < 2
+    ) {
       // 0.1.8.x was the first server version to support rooms.
-      console.warn("[CloudLink] Server is too old! Must be at least 0.1.8.x to support room linking/unlinking.");
+      console.warn(
+        "[CloudLink] Server is too old! Must be at least 0.1.8.x to support room linking/unlinking."
+      );
       return;
     }
 
@@ -371,7 +384,10 @@
     try {
       outgoing = JSON.stringify(message);
     } catch (SyntaxError) {
-      console.warn("[CloudLink] Failed to send a packet, invalid syntax:", message);
+      console.warn(
+        "[CloudLink] Failed to send a packet, invalid syntax:",
+        message
+      );
       return;
     }
 
@@ -393,12 +409,13 @@
           versionNumber: version.versionNumber,
         },
       },
-      listener: "handshake_cfg"
+      listener: "handshake_cfg",
     });
     clVars.handshakeAttempted = true;
   }
 
   // Compare the version string of the server to known compatible variants to configure clVars.linkState.identifiedProtocol.
+  // eslint-disable-next-line require-await
   async function setServerVersion(version) {
     console.log(`[CloudLink] Server version: ${String(version)}`);
     clVars.server_version = version;
@@ -413,47 +430,56 @@
       "S2.2": 0, // 0.1.5
       "0.1.": 0, // 0.1.5 or legacy
       "S2.": 0, // Legacy
-      "S1.": -1 // Obsolete
+      "S1.": -1, // Obsolete
     };
 
     for (const [key, value] of Object.entries(versions)) {
       if (version.includes(key)) {
         if (clVars.linkState.identifiedProtocol < value) {
-
           // Disconnect if protcol is too old
           if (value == -1) {
-            console.warn(`[CloudLink] Server is too old to enable leagacy support. Disconnecting.`);
+            console.warn(
+              `[CloudLink] Server is too old to enable leagacy support. Disconnecting.`
+            );
             return clVars.socket.close(1000, "");
           }
-          
+
           // Set the identified protocol variant
           clVars.linkState.identifiedProtocol = value;
         }
       }
-    };
+    }
 
     // Log configured spec version
-    console.log(`[CloudLink] Configured protocol spec to v${clVars.linkState.identifiedProtocol}.`);
+    console.log(
+      `[CloudLink] Configured protocol spec to v${clVars.linkState.identifiedProtocol}.`
+    );
 
     // Fix timing bug
     clVars.linkState.status = 2;
 
     // Fire event hats (only one not broken)
-    runtime.startHats('cloudlink_onConnect');
+    runtime.startHats("cloudlink_onConnect");
 
     // Don't nag user if they already trusted this server
     if (clVars.currentServerUrl === clVars.lastServerUrl) return;
 
     // Ask user if they wish to stay connected if the server is unsupported
-    if ((clVars.linkState.identifiedProtocol < 4) && (!confirm(
-      `You have connected to an old CloudLink server, running version ${clVars.server_version}.\n\nFor your security and privacy, we recommend you disconnect from this server and connect to an up-to-date server.\n\nClick/tap \"OK\" to stay connected.`
-    ))) {
+    if (
+      clVars.linkState.identifiedProtocol < 4 &&
+      !confirm(
+        `You have connected to an old CloudLink server, running version ${clVars.server_version}.\n\nFor your security and privacy, we recommend you disconnect from this server and connect to an up-to-date server.\n\nClick/tap "OK" to stay connected.`
+      )
+    ) {
       // Close the connection if they choose "Cancel"
       clVars.linkState.isAttemptingGracefulDisconnect = true;
-      clVars.socket.close(1000, "Client going away (legacy server rejected by end user)");
+      clVars.socket.close(
+        1000,
+        "Client going away (legacy server rejected by end user)"
+      );
       return;
     }
-    
+
     // Don't nag user the next time they connect to this server
     clVars.lastServerUrl = clVars.currentServerUrl;
   }
@@ -463,15 +489,21 @@
     // Parse the message JSON
     let packet = {};
     try {
-      packet = JSON.parse(data)
+      packet = JSON.parse(data);
     } catch (SyntaxError) {
-      console.error("[CloudLink] Incoming message parse failure! Is this really a CloudLink server?", data);
+      console.error(
+        "[CloudLink] Incoming message parse failure! Is this really a CloudLink server?",
+        data
+      );
       return;
-    };
+    }
 
     // Handle packet commands
-    if (!packet.hasOwnProperty("cmd")) {
-      console.error("[CloudLink] Incoming message read failure! This message doesn't contain the required \"cmd\" key. Is this really a CloudLink server?", packet);
+    if (!Object.prototype.hasOwnProperty.call(packet, "cmd")) {
+      console.error(
+        '[CloudLink] Incoming message read failure! This message doesn\'t contain the required "cmd" key. Is this really a CloudLink server?',
+        packet
+      );
       return;
     }
     console.log("[CloudLink] RX:", packet);
@@ -512,7 +544,7 @@
 
       case "direct":
         // Handle events from older server versions
-        if (packet.val.hasOwnProperty("cmd")) {
+        if (Object.prototype.hasOwnProperty.call(packet.val, "cmd")) {
           switch (packet.val.cmd) {
             // Server 0.1.5 (at least)
             case "vers":
@@ -522,7 +554,9 @@
 
             // Server 0.1.7 (at least)
             case "motd":
-              console.log(`[CloudLink] Message of the day: \"${packet.val.val}\"`);
+              console.log(
+                `[CloudLink] Message of the day: "${packet.val.val}"`
+              );
               clVars.motd = packet.val.val;
               return;
           }
@@ -544,7 +578,9 @@
         // Store direct value
         // Protocol v0 (0.1.5 and legacy) don't implement status codes.
         if (clVars.linkState.identifiedProtocol == 0) {
-          console.warn("[CloudLink] Received a statuscode message while using protocol v0. This event shouldn't happen. It's likely that this server is modified (did MikeDEV overlook some unexpected behavior?).");
+          console.warn(
+            "[CloudLink] Received a statuscode message while using protocol v0. This event shouldn't happen. It's likely that this server is modified (did MikeDEV overlook some unexpected behavior?)."
+          );
           return;
         }
 
@@ -556,30 +592,32 @@
         // Protocol v2 (0.1.8.x) uses "code" instead.
         // Protocol v3-v4 (0.1.9.x - latest, 0.2.0) adds "code_id" to the payload. Ignored by Scratch clients.
         else {
-
           // Handle setup listeners
-          if (packet.hasOwnProperty("listener")) {
+          if (Object.prototype.hasOwnProperty.call(packet, "listener")) {
             switch (packet.listener) {
               case "username_cfg":
-
                 // Username accepted
                 if (packet.code.includes("I:100")) {
                   clVars.myUserObject = packet.val;
                   clVars.username.value = packet.val.username;
                   clVars.username.accepted = true;
-                  console.log(`[CloudLink] Username has been set to \"${clVars.username.value}\" successfully!`);
+                  console.log(
+                    `[CloudLink] Username has been set to "${clVars.username.value}" successfully!`
+                  );
 
-                // Username rejected / error
+                  // Username rejected / error
                 } else {
-                  console.log(`[CloudLink] Username rejected by the server! Error code ${packet.code}.}`);
+                  console.log(
+                    `[CloudLink] Username rejected by the server! Error code ${packet.code}.}`
+                  );
                 }
                 return;
-              
+
               case "handshake_cfg":
                 // Prevent handshake responses being stored in the statuscode variables
                 console.log("[CloudLink] Server responded to our handshake!");
                 return;
-              
+
               case "link":
                 // Room link accepted
                 if (!clVars.rooms.isAttemptingLink) return;
@@ -588,12 +626,14 @@
                   clVars.rooms.isLinked = true;
                   console.log("[CloudLink] Room linked successfully!");
 
-                // Room link rejected / error
+                  // Room link rejected / error
                 } else {
-                  console.log(`[CloudLink] Room link rejected! Error code ${packet.code}.}`);
+                  console.log(
+                    `[CloudLink] Room link rejected! Error code ${packet.code}.}`
+                  );
                 }
                 return;
-              
+
               case "unlink":
                 // Room unlink accepted
                 if (!clVars.rooms.isAttemptingUnlink) return;
@@ -602,9 +642,11 @@
                   clVars.rooms.isLinked = false;
                   console.log("[CloudLink] Room unlinked successfully!");
 
-                // Room link rejected / error
+                  // Room link rejected / error
                 } else {
-                  console.log(`[CloudLink] Room unlink rejected! Error code ${packet.code}.}`);
+                  console.log(
+                    `[CloudLink] Room unlink rejected! Error code ${packet.code}.}`
+                  );
                 }
                 return;
             }
@@ -623,21 +665,25 @@
       case "ulist":
         // Protocol v0-v1 (0.1.5 and legacy - 0.1.7) use a semicolon (;) separated string for the userlist.
         if (
-          (clVars.linkState.identifiedProtocol == 0)
-          || 
-          (clVars.linkState.identifiedProtocol == 1)
+          clVars.linkState.identifiedProtocol == 0 ||
+          clVars.linkState.identifiedProtocol == 1
         ) {
           // Split the username list string
-          clVars.ulist = String(packet.val).split(';');
+          clVars.ulist = String(packet.val).split(";");
 
           // Get rid of blank entry at the end of the list
-          clVars.ulist.pop(clVars.ulist.length);
+          clVars.ulist.pop();
 
           // Check if username has been set (since older servers don't implement statuscodes or listeners)
-          if ((clVars.username.attempted) && (clVars.ulist.includes(clVars.username.temp))) {
+          if (
+            clVars.username.attempted &&
+            clVars.ulist.includes(clVars.username.temp)
+          ) {
             clVars.username.value = clVars.username.temp;
             clVars.username.accepted = true;
-            console.log(`[CloudLink] Username has been set to \"${clVars.username.value}\" successfully!`);
+            console.log(
+              `[CloudLink] Username has been set to "${clVars.username.value}" successfully!`
+            );
           }
         }
 
@@ -649,23 +695,27 @@
         // Protocol v3-v4 (0.1.9.x - latest, 0.2.0) uses "mode" to add/set/remove entries to the userlist.
         else {
           // Check for "mode" key
-          if (!packet.hasOwnProperty("mode")) {
-            console.warn("[CloudLink] Userlist message did not specify \"mode\" while running in protocol mode 3 or 4.");
+          if (!Object.prototype.hasOwnProperty.call(packet, "mode")) {
+            console.warn(
+              '[CloudLink] Userlist message did not specify "mode" while running in protocol mode 3 or 4.'
+            );
             return;
-          };
+          }
           // Handle methods
           switch (packet.mode) {
-            case 'set':
+            case "set":
               clVars.ulist = packet.val;
               break;
-            case 'add':
+            case "add":
               clVars.ulist.push(packet.val);
               break;
-            case 'remove':
+            case "remove":
               clVars.ulist.slice(clVars.ulist.indexOf(packet.val), 1);
               break;
             default:
-              console.warn(`[CloudLink] Unrecognised userlist mode: \"${packet.mode}\".`);
+              console.warn(
+                `[CloudLink] Unrecognised userlist mode: "${packet.mode}".`
+              );
               break;
           }
         }
@@ -680,27 +730,28 @@
 
       case "client_ip":
         console.log(`[CloudLink] Client IP address: ${packet.val}`);
-        console.warn("[CloudLink] This server has relayed your identified IP address to you. Under normal circumstances, this will be erased server-side when you disconnect, but you should still be careful. Unless you trust this server, it is not recommended to send login credentials or personal info.");
+        console.warn(
+          "[CloudLink] This server has relayed your identified IP address to you. Under normal circumstances, this will be erased server-side when you disconnect, but you should still be careful. Unless you trust this server, it is not recommended to send login credentials or personal info."
+        );
         clVars.client_ip = packet.val;
         break;
 
       case "motd":
-        console.log(`[CloudLink] Message of the day: \"${packet.val}\"`);
+        console.log(`[CloudLink] Message of the day: "${packet.val}"`);
         clVars.motd = packet.val;
         break;
 
       default:
-        console.warn(`[CloudLink] Unrecognised command: \"${packet.cmd}\".`);
+        console.warn(`[CloudLink] Unrecognised command: "${packet.cmd}".`);
         return;
     }
 
     // Handle listeners
-    if (packet.hasOwnProperty("listener")) {
+    if (Object.prototype.hasOwnProperty.call(packet, "listener")) {
       if (clVars.listeners.current.includes(String(packet.listener))) {
-
         // Remove the listener from the currently listening list
         clVars.listeners.current.splice(
-          clVars.listeners.current.indexOf(String(packet.listener)), 
+          clVars.listeners.current.indexOf(String(packet.listener)),
           1
         );
 
@@ -717,7 +768,9 @@
   // Basic netcode needed to make the extension work
   async function newClient(url) {
     if (!(await Scratch.canFetch(url))) {
-      console.warn("[CloudLink] Did not get permission to connect, aborting...");
+      console.warn(
+        "[CloudLink] Did not get permission to connect, aborting..."
+      );
       return;
     }
 
@@ -728,6 +781,7 @@
     // Establish a connection to the server
     console.log("[CloudLink] Connecting to server:", url);
     try {
+      // eslint-disable-next-line extension/check-can-fetch
       clVars.socket = new WebSocket(url);
     } catch (e) {
       console.warn("[CloudLink] An exception has occurred:", e);
@@ -737,13 +791,15 @@
     // Bind connection established event
     clVars.socket.onopen = function (event) {
       clVars.currentServerUrl = url;
-      
+
       // Set the link state to connected.
       console.log("[CloudLink] Connected.");
 
       // If a server_version message hasn't been received in over half a second, try to broadcast a handshake
-      clVars.handshakeTimeout = window.setTimeout(function() {
-        console.log("[CloudLink] Hmm... This server hasn't sent us it's server info. Going to attempt a handshake.");
+      clVars.handshakeTimeout = window.setTimeout(function () {
+        console.log(
+          "[CloudLink] Hmm... This server hasn't sent us it's server info. Going to attempt a handshake."
+        );
         sendHandshake();
       }, 500);
 
@@ -767,14 +823,21 @@
           break;
 
         case 2: // Was already connected
-          if (event.wasClean || clVars.linkState.isAttemptingGracefulDisconnect) {
+          if (
+            event.wasClean ||
+            clVars.linkState.isAttemptingGracefulDisconnect
+          ) {
             // Set the link state to graceful disconnect.
-            console.log(`[CloudLink] Disconnected (${event.code} ${event.reason}).`);
+            console.log(
+              `[CloudLink] Disconnected (${event.code} ${event.reason}).`
+            );
             clVars.linkState.status = 3;
             clVars.linkState.disconnectType = 0;
           } else {
             // Set the link state to ungraceful disconnect.
-            console.log(`[CloudLink] Lost connection (${event.code} ${event.reason}).`);
+            console.log(
+              `[CloudLink] Lost connection (${event.code} ${event.reason}).`
+            );
             clVars.linkState.status = 4;
             clVars.linkState.disconnectType = 2;
           }
@@ -785,59 +848,39 @@
       resetOnClose();
 
       // Run all onClose event blocks
-      runtime.startHats('cloudlink_onClose');
+      runtime.startHats("cloudlink_onClose");
       // Return promise (during setup)
       return;
-    }
-  }
-
-  // GET the serverList
-  try {
-    Scratch.fetch(
-      "https://raw.githubusercontent.com/MikeDev101/cloudlink/master/serverlist.json"
-    )
-      .then((response) => {
-        return response.text();
-      })
-      .then((data) => {
-        clVars.serverList = JSON.parse(data);
-      })
-      .catch((err) => {
-        console.log("[CloudLink] An error has occurred while parsing the public server list:", err);
-        clVars.serverList = {};
-      });
-  } catch (err) {
-    console.log("[CloudLink] An error has occurred while fetching the public server list:", err);
-    clVars.serverList = {};
+    };
   }
 
   // Declare the CloudLink library.
   class CloudLink {
     getInfo() {
       return {
-        id: 'cloudlink',
-        name: 'CloudLink',
+        id: "cloudlink",
+        // eslint-disable-next-line extension/should-translate
+        name: "CloudLink V4",
         blockIconURI: cl_block,
         menuIconURI: cl_icon,
         docsURI: "https://github.com/MikeDev101/cloudlink/wiki/Scratch-Client",
         blocks: [
-
           {
             opcode: "returnGlobalData",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("global data")
+            text: Scratch.translate("global data"),
           },
 
           {
             opcode: "returnPrivateData",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("private data")
+            text: Scratch.translate("private data"),
           },
 
           {
             opcode: "returnDirectData",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("direct data")
+            text: Scratch.translate("direct data"),
           },
 
           "---",
@@ -845,13 +888,13 @@
           {
             opcode: "returnLinkData",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("link status")
+            text: Scratch.translate("link status"),
           },
 
           {
             opcode: "returnStatusCode",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("status code")
+            text: Scratch.translate("status code"),
           },
 
           "---",
@@ -859,20 +902,22 @@
           {
             opcode: "returnUserListData",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("usernames")
+            text: Scratch.translate("usernames"),
           },
 
           {
             opcode: "returnUsernameDataNew",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("my username")
+            text: Scratch.translate("my username"),
           },
 
           {
             opcode: "returnUsernameData",
             blockType: Scratch.BlockType.REPORTER,
             hideFromPalette: clVars.hideCLDeprecatedBlocks,
-            text: Scratch.translate("(OLD - DO NOT USE IN NEW PROJECTS) my username")
+            text: Scratch.translate(
+              "(OLD - DO NOT USE IN NEW PROJECTS) my username"
+            ),
           },
 
           "---",
@@ -880,25 +925,25 @@
           {
             opcode: "returnVersionData",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("extension version")
+            text: Scratch.translate("extension version"),
           },
 
           {
             opcode: "returnServerVersion",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("server version")
+            text: Scratch.translate("server version"),
           },
 
           {
             opcode: "returnServerList",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("server list")
+            text: Scratch.translate("server list"),
           },
 
           {
             opcode: "returnMOTD",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("server MOTD")
+            text: Scratch.translate("server MOTD"),
           },
 
           "---",
@@ -906,13 +951,13 @@
           {
             opcode: "returnClientIP",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("my IP address")
+            text: Scratch.translate("my IP address"),
           },
 
           {
             opcode: "returnUserObject",
             blockType: Scratch.BlockType.REPORTER,
-            text: Scratch.translate("my user object")
+            text: Scratch.translate("my user object"),
           },
 
           "---",
@@ -986,11 +1031,12 @@
             arguments: {
               PATH: {
                 type: Scratch.ArgumentType.STRING,
-                defaultValue: 'fruit/apples',
+                defaultValue: "fruit/apples",
               },
               JSON_STRING: {
                 type: Scratch.ArgumentType.STRING,
-                defaultValue: '{"fruit": {"apples": 2, "bananas": 3}, "total_fruit": 5}',
+                defaultValue:
+                  '{"fruit": {"apples": 2, "bananas": 3}, "total_fruit": 5}',
               },
             },
           },
@@ -999,7 +1045,7 @@
             opcode: "getFromJSONArray",
             blockType: Scratch.BlockType.REPORTER,
             hideFromPalette: clVars.hideCLDeprecatedBlocks,
-            text: Scratch.translate('[NUM] from JSON array [ARRAY]'),
+            text: Scratch.translate("[NUM] from JSON array [ARRAY]"),
             arguments: {
               NUM: {
                 type: Scratch.ArgumentType.NUMBER,
@@ -1008,8 +1054,8 @@
               ARRAY: {
                 type: Scratch.ArgumentType.STRING,
                 defaultValue: '["foo","bar"]',
-              }
-            }
+              },
+            },
           },
 
           {
@@ -1033,11 +1079,11 @@
             arguments: {
               JSON_STRING: {
                 type: Scratch.ArgumentType.STRING,
-                defaultValue: '{"fruit": {"apples": 2, "bananas": 3}, "total_fruit": 5}',
+                defaultValue:
+                  '{"fruit": {"apples": 2, "bananas": 3}, "total_fruit": 5}',
               },
             },
           },
-
 
           "---",
 
@@ -1058,7 +1104,9 @@
             opcode: "requestURL",
             blockType: Scratch.BlockType.REPORTER,
             hideFromPalette: clVars.hideCLDeprecatedBlocks,
-            text: Scratch.translate("send request with method [method] for URL [url] with data [data] and headers [headers]"),
+            text: Scratch.translate(
+              "send request with method [method] for URL [url] with data [data] and headers [headers]"
+            ),
             arguments: {
               method: {
                 type: Scratch.ArgumentType.STRING,
@@ -1100,7 +1148,9 @@
           {
             opcode: "onListener",
             blockType: Scratch.BlockType.HAT,
-            text: Scratch.translate("when I receive new message with listener [ID]"),
+            text: Scratch.translate(
+              "when I receive new message with listener [ID]"
+            ),
             isEdgeActivated: true,
             arguments: {
               ID: {
@@ -1237,9 +1287,9 @@
             arguments: {
               IP: {
                 type: Scratch.ArgumentType.STRING,
-                defaultValue: "ws://127.0.0.1:3000/",
-              }
-            }
+                defaultValue: "wss://cl.mikedev101.cc/",
+              },
+            },
           },
 
           {
@@ -1249,15 +1299,15 @@
             arguments: {
               ID: {
                 type: Scratch.ArgumentType.NUMBER,
-                defaultValue: 1,
-              }
-            }
+                defaultValue: "7",
+              },
+            },
           },
 
           {
             opcode: "closeSocket",
             blockType: Scratch.BlockType.COMMAND,
-            text: Scratch.translate("disconnect")
+            text: Scratch.translate("disconnect"),
           },
 
           "---",
@@ -1286,21 +1336,20 @@
                 defaultValue: "example-listener",
               },
             },
-
           },
 
           "---",
 
           {
-            opcode: 'linkToRooms',
+            opcode: "linkToRooms",
             blockType: Scratch.BlockType.COMMAND,
             text: Scratch.translate("link to room(s) [ROOMS]"),
             arguments: {
               ROOMS: {
                 type: Scratch.ArgumentType.STRING,
-                defaultValue: Scratch.translate('["test"]'),
+                defaultValue: JSON.stringify(["test"]),
               },
-            }
+            },
           },
 
           {
@@ -1310,7 +1359,7 @@
             arguments: {
               ROOMS: {
                 type: Scratch.ArgumentType.STRING,
-                defaultValue: Scratch.translate('["test"]'),
+                defaultValue: JSON.stringify(["test"]),
               },
             },
           },
@@ -1370,7 +1419,9 @@
           {
             opcode: "sendPDataAsVar",
             blockType: Scratch.BlockType.COMMAND,
-            text: Scratch.translate("send variable [VAR] to [ID] with data [DATA]"),
+            text: Scratch.translate(
+              "send variable [VAR] to [ID] with data [DATA]"
+            ),
             arguments: {
               DATA: {
                 type: Scratch.ArgumentType.STRING,
@@ -1485,7 +1536,7 @@
                 menu: "allmenu",
                 defaultValue: "All data",
               },
-            }
+            },
           },
 
           "---",
@@ -1505,45 +1556,71 @@
           },
 
           "---",
-
         ],
         menus: {
           datamenu: {
             items: [
-              {text: Scratch.translate('Global data'), value: 'Global data'},
-              {text: Scratch.translate('Private data'), value: 'Private data'},
-              {text: Scratch.translate('Direct data'), value: 'Direct data'},
-              {text: Scratch.translate('Status code'), value: 'Status code'}
-            ]
+              { text: Scratch.translate("Global data"), value: "Global data" },
+              {
+                text: Scratch.translate("Private data"),
+                value: "Private data",
+              },
+              { text: Scratch.translate("Direct data"), value: "Direct data" },
+              { text: Scratch.translate("Status code"), value: "Status code" },
+            ],
           },
           varmenu: {
             items: [
-              {text: Scratch.translate('Global variables'), value: "Global variables"},
-              {text: Scratch.translate('Private variables'), value: "Private variables"}
-            ]
+              {
+                text: Scratch.translate("Global variables"),
+                value: "Global variables",
+              },
+              {
+                text: Scratch.translate("Private variables"),
+                value: "Private variables",
+              },
+            ],
           },
           allmenu: {
             items: [
-              {text: Scratch.translate('Global data'), value: 'Global data'},
-              {text: Scratch.translate('Private data'), value: 'Private data'},
-              {text: Scratch.translate('Direct data'), value: 'Direct data'},
-              {text: Scratch.translate('Status code'), value: 'Status code'},
-              {text: Scratch.translate("Global variables"), value: "Global variables"},
-              {text: Scratch.translate("Private variables"), value: "Private variables"},
-              {text: Scratch.translate("All data"), value: "All data"}
-              ]
+              { text: Scratch.translate("Global data"), value: "Global data" },
+              {
+                text: Scratch.translate("Private data"),
+                value: "Private data",
+              },
+              { text: Scratch.translate("Direct data"), value: "Direct data" },
+              { text: Scratch.translate("Status code"), value: "Status code" },
+              {
+                text: Scratch.translate("Global variables"),
+                value: "Global variables",
+              },
+              {
+                text: Scratch.translate("Private variables"),
+                value: "Private variables",
+              },
+              { text: Scratch.translate("All data"), value: "All data" },
+            ],
           },
           almostallmenu: {
             items: [
-              {text: Scratch.translate('Global data'), value: 'Global data'},
-              {text: Scratch.translate('Private data'), value: 'Private data'},
-              {text: Scratch.translate('Direct data'), value: 'Direct data'},
-              {text: Scratch.translate('Status code'), value: 'Status code'},
-              {text: Scratch.translate("Global variables"), value: "Global variables"},
-              {text: Scratch.translate("Private variables"), value: "Private variables"}
-              ]
+              { text: Scratch.translate("Global data"), value: "Global data" },
+              {
+                text: Scratch.translate("Private data"),
+                value: "Private data",
+              },
+              { text: Scratch.translate("Direct data"), value: "Direct data" },
+              { text: Scratch.translate("Status code"), value: "Status code" },
+              {
+                text: Scratch.translate("Global variables"),
+                value: "Global variables",
+              },
+              {
+                text: Scratch.translate("Private variables"),
+                value: "Private variables",
+              },
+            ],
           },
-        }
+        },
       };
     }
 
@@ -1638,7 +1715,12 @@
     // Reporter - Returns data for a specific listener ID.
     // ID - String (listener ID)
     returnListenerData(args) {
-      if (!clVars.listeners.varStates.hasOwnProperty(String(args.ID))) {
+      if (
+        !Object.prototype.hasOwnProperty.call(
+          clVars.listeners.varStates,
+          String(args.ID)
+        )
+      ) {
         console.warn(`[CloudLink] Listener ID ${args.ID} does not exist!`);
         return "";
       }
@@ -1649,19 +1731,19 @@
     // TYPE - String (menu allmenu)
     readQueueSize(args) {
       switch (args.TYPE) {
-        case 'Global data':
+        case "Global data":
           return clVars.gmsg.queue.length;
-        case 'Private data':
+        case "Private data":
           return clVars.pmsg.queue.length;
-        case 'Direct data':
+        case "Direct data":
           return clVars.direct.queue.length;
-        case 'Status code':
+        case "Status code":
           return clVars.statuscode.queue.length;
-        case 'Global variables':
+        case "Global variables":
           return clVars.gvar.queue.length;
-        case 'Private variables':
+        case "Private variables":
           return clVars.pvar.queue.length;
-        case 'All data':
+        case "All data":
           return (
             clVars.gmsg.queue.length +
             clVars.pmsg.queue.length +
@@ -1677,26 +1759,26 @@
     // TYPE - String (menu allmenu)
     readQueueData(args) {
       switch (args.TYPE) {
-        case 'Global data':
+        case "Global data":
           return makeValueScratchSafe(clVars.gmsg.queue);
-        case 'Private data':
+        case "Private data":
           return makeValueScratchSafe(clVars.pmsg.queue);
-        case 'Direct data':
+        case "Direct data":
           return makeValueScratchSafe(clVars.direct.queue);
-        case 'Status code':
+        case "Status code":
           return makeValueScratchSafe(clVars.statuscode.queue);
-        case 'Global variables':
+        case "Global variables":
           return makeValueScratchSafe(clVars.gvar.queue);
-        case 'Private variables':
+        case "Private variables":
           return makeValueScratchSafe(clVars.pvar.queue);
-        case 'All data':
+        case "All data":
           return makeValueScratchSafe({
             gmsg: clVars.gmsg.queue,
             pmsg: clVars.pmsg.queue,
             direct: clVars.direct.queue,
             statuscode: clVars.statuscode.queue,
             gvar: clVars.gvar.queue,
-            pvar: clVars.pvar.queue
+            pvar: clVars.pvar.queue,
           });
       }
     }
@@ -1705,42 +1787,58 @@
     // TYPE - String (menu varmenu), VAR - String (variable name)
     returnVarData(args) {
       switch (args.TYPE) {
-      case 'Global variables':
-        if (!clVars.gvar.varStates.hasOwnProperty(String(args.VAR))) {
-          console.warn(`[CloudLink] Global variable ${args.VAR} does not exist!`);
-          return "";
-        }
-        return clVars.gvar.varStates[String(args.VAR)].varState;
-      case 'Private variables':
-        if (!clVars.pvar.varStates.hasOwnProperty(String(args.VAR))) {
-          console.warn(`[CloudLink] Private variable ${args.VAR} does not exist!`);
-          return "";
-        }
-        return clVars.pvar.varStates[String(args.VAR)].varState;
+        case "Global variables":
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.gvar.varStates,
+              String(args.VAR)
+            )
+          ) {
+            console.warn(
+              `[CloudLink] Global variable ${args.VAR} does not exist!`
+            );
+            return "";
+          }
+          return clVars.gvar.varStates[String(args.VAR)].varState;
+        case "Private variables":
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.pvar.varStates,
+              String(args.VAR)
+            )
+          ) {
+            console.warn(
+              `[CloudLink] Private variable ${args.VAR} does not exist!`
+            );
+            return "";
+          }
+          return clVars.pvar.varStates[String(args.VAR)].varState;
       }
     }
 
     // Reporter - Gets a JSON key value from a JSON string.
     // PATH - String, JSON_STRING - String
-    parseJSON(args) { 
+    parseJSON(args) {
       try {
-        const path = args.PATH.toString().split('/').map(prop => decodeURIComponent(prop));
-        if (path[0] === '') path.splice(0, 1);
-        if (path[path.length - 1] === '') path.splice(-1, 1);
+        const path = args.PATH.toString()
+          .split("/")
+          .map((prop) => decodeURIComponent(prop));
+        if (path[0] === "") path.splice(0, 1);
+        if (path[path.length - 1] === "") path.splice(-1, 1);
         let json;
         try {
-          json = JSON.parse(' ' + args.JSON_STRING);
+          json = JSON.parse(" " + args.JSON_STRING);
         } catch (e) {
           return e.message;
-        };
-        path.forEach(prop => json = json[prop]);
-        if (json === null) return 'null';
-        else if (json === undefined) return '';
-        else if (typeof json === 'object') return JSON.stringify(json);
+        }
+        path.forEach((prop) => (json = json[prop]));
+        if (json === null) return "null";
+        else if (json === undefined) return "";
+        else if (typeof json === "object") return JSON.stringify(json);
         else return json.toString();
       } catch (err) {
-        return '';
-      };
+        return "";
+      }
     }
 
     // Reporter - Returns an entry from a JSON array (0-based).
@@ -1751,11 +1849,11 @@
         return "";
       } else {
         let data = json_array[args.NUM];
-  
-        if (typeof (data) == "object") {
+
+        if (typeof data == "object") {
           data = JSON.stringify(data); // Make the JSON safe for Scratch
         }
-  
+
         return data;
       }
     }
@@ -1763,9 +1861,9 @@
     // Reporter - Returns a RESTful GET promise.
     // url - String
     fetchURL(args) {
-      return Scratch.fetch(args.url, {method: "GET"})
-        .then(response => response.text())
-        .catch(error => {
+      return Scratch.fetch(args.url, { method: "GET" })
+        .then((response) => response.text())
+        .catch((error) => {
           console.warn(`[CloudLink] Fetch error: ${error}`);
         });
     }
@@ -1776,25 +1874,25 @@
       if (args.method == "GET" || args.method == "HEAD") {
         return Scratch.fetch(args.url, {
           method: args.method,
-          headers: JSON.parse(args.headers)
+          headers: JSON.parse(args.headers),
         })
-          .then(response => response.text())
-          .catch(error => {
+          .then((response) => response.text())
+          .catch((error) => {
             console.warn(`[CloudLink] Request error: ${error}`);
           });
       } else {
         return Scratch.fetch(args.url, {
           method: args.method,
           headers: JSON.parse(args.headers),
-          body: JSON.parse(args.data)
+          body: JSON.parse(args.data),
         })
-          .then(response => response.text())
-          .catch(error => {
+          .then((response) => response.text())
+          .catch((error) => {
             console.warn(`[CloudLink] Request error: ${error}`);
           });
       }
     }
-    
+
     // Event
     // ID - String (listener)
     onListener(args) {
@@ -1803,7 +1901,13 @@
       if (clVars.linkState.status != 2) return false;
 
       // Listener must exist
-      if (!clVars.listeners.varStates.hasOwnProperty(args.ID)) return false;
+      if (
+        !Object.prototype.hasOwnProperty.call(
+          clVars.listeners.varStates,
+          args.ID
+        )
+      )
+        return false;
 
       // Run event
       if (clVars.listeners.varStates[args.ID].eventHatTick) {
@@ -1822,42 +1926,42 @@
 
       // Run event
       switch (args.TYPE) {
-        case 'Global data':
+        case "Global data":
           if (clVars.gmsg.eventHatTick) {
             clVars.gmsg.eventHatTick = false;
             return true;
           }
           break;
 
-        case 'Private data':
+        case "Private data":
           if (clVars.pmsg.eventHatTick) {
             clVars.pmsg.eventHatTick = false;
             return true;
           }
           break;
 
-        case 'Direct data':
+        case "Direct data":
           if (clVars.direct.eventHatTick) {
             clVars.direct.eventHatTick = false;
             return true;
           }
           break;
 
-        case 'Status code':
+        case "Status code":
           if (clVars.statuscode.eventHatTick) {
             clVars.statuscode.eventHatTick = false;
             return true;
           }
           break;
 
-        case 'Global variables':
+        case "Global variables":
           if (clVars.gvar.eventHatTick) {
             clVars.gvar.eventHatTick = false;
             return true;
           }
           break;
 
-        case 'Private variables':
+        case "Private variables":
           if (clVars.pvar.eventHatTick) {
             clVars.pvar.eventHatTick = false;
             return true;
@@ -1876,10 +1980,15 @@
 
       // Run event
       switch (args.TYPE) {
-        case 'Global variables':
-
+        case "Global variables":
           // Variable must exist
-          if (!clVars.gvar.varStates.hasOwnProperty(String(args.VAR))) break;
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.gvar.varStates,
+              String(args.VAR)
+            )
+          )
+            break;
           if (clVars.gvar.varStates[String(args.VAR)].eventHatTick) {
             clVars.gvar.varStates[String(args.VAR)].eventHatTick = false;
             return true;
@@ -1887,10 +1996,15 @@
 
           break;
 
-        case 'Private variables':
-
+        case "Private variables":
           // Variable must exist
-          if (!clVars.pvar.varStates.hasOwnProperty(String(args.VAR))) break;
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.pvar.varStates,
+              String(args.VAR)
+            )
+          )
+            break;
           if (clVars.pvar.varStates[String(args.VAR)].eventHatTick) {
             clVars.pvar.varStates[String(args.VAR)].eventHatTick = false;
             return true;
@@ -1904,61 +2018,64 @@
     // Reporter - Returns a JSON-ified value.
     // toBeJSONified - String
     makeJSON(args) {
-      if (typeof(args.toBeJSONified) == "string") {
+      if (typeof args.toBeJSONified == "string") {
         try {
           JSON.parse(args.toBeJSONified);
           return String(args.toBeJSONified);
-        } catch(err) {
+        } catch (err) {
           return "Not JSON!";
         }
-      } else if (typeof(args.toBeJSONified) == "object") {
+      } else if (typeof args.toBeJSONified == "object") {
         return JSON.stringify(args.toBeJSONified);
       } else {
         return "Not JSON!";
-      };
+      }
     }
 
     // Boolean - Returns true if connected.
     getComState() {
-      return ((clVars.linkState.status == 2) && (clVars.socket != null));
+      return clVars.linkState.status == 2 && clVars.socket != null;
     }
 
     // Boolean - Returns true if linked to rooms (other than "default")
     getRoomState() {
-      return ((clVars.socket != null) && (clVars.rooms.isLinked));
+      return clVars.socket != null && clVars.rooms.isLinked;
     }
 
     // Boolean - Returns true if the connection was dropped.
     getComLostConnectionState() {
-      return ((clVars.linkState.status == 4) && (clVars.linkState.disconnectType == 2));
+      return (
+        clVars.linkState.status == 4 && clVars.linkState.disconnectType == 2
+      );
     }
 
     // Boolean - Returns true if the client failed to establish a connection.
     getComFailedConnectionState() {
-      return ((clVars.linkState.status == 4) && (clVars.linkState.disconnectType == 1));
+      return (
+        clVars.linkState.status == 4 && clVars.linkState.disconnectType == 1
+      );
     }
 
     // Boolean - Returns true if the username was set successfully.
     getUsernameState() {
-      return ((clVars.socket != null) && (clVars.username.accepted));
+      return clVars.socket != null && clVars.username.accepted;
     }
 
     // Boolean - Returns true if there is new gmsg/pmsg/direct/statuscode data.
     // TYPE - String (menu datamenu)
     returnIsNewData(args) {
-
       // Must be connected
       if (clVars.socket == null) return false;
 
       // Run event
       switch (args.TYPE) {
-        case 'Global data':
+        case "Global data":
           return clVars.gmsg.hasNew;
-        case 'Private data':
+        case "Private data":
           return clVars.pmsg.hasNew;
-        case 'Direct data':
+        case "Direct data":
           return clVars.direct.hasNew;
-        case 'Status code':
+        case "Status code":
           return clVars.statuscode.hasNew;
       }
     }
@@ -1967,15 +2084,29 @@
     // TYPE - String (menu varmenu), VAR - String (variable name)
     returnIsNewVarData(args) {
       switch (args.TYPE) {
-        case 'Global variables':
-          if (!clVars.gvar.varStates.hasOwnProperty(String(args.VAR))) {
-            console.warn(`[CloudLink] Global variable ${args.VAR} does not exist!`);
+        case "Global variables":
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.gvar.varStates,
+              String(args.VAR)
+            )
+          ) {
+            console.warn(
+              `[CloudLink] Global variable ${args.VAR} does not exist!`
+            );
             return false;
           }
           return clVars.gvar.varStates[String(args.ID)].hasNew;
-        case 'Private variables':
-          if (!clVars.pvar.varStates.hasOwnProperty(String(args.VAR))) {
-            console.warn(`[CloudLink] Private variable ${args.VAR} does not exist!`);
+        case "Private variables":
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.pvar.varStates,
+              String(args.VAR)
+            )
+          ) {
+            console.warn(
+              `[CloudLink] Private variable ${args.VAR} does not exist!`
+            );
             return false;
           }
           return clVars.pvar.varStates[String(args.ID)].hasNew;
@@ -1985,7 +2116,12 @@
     // Boolean - Returns true if a listener has a new value.
     // ID - String (listener ID)
     returnIsNewListener(args) {
-      if (!clVars.listeners.varStates.hasOwnProperty(String(args.ID))) {
+      if (
+        !Object.prototype.hasOwnProperty.call(
+          clVars.listeners.varStates,
+          String(args.ID)
+        )
+      ) {
         console.warn(`[CloudLink] Listener ID ${args.ID} does not exist!`);
         return false;
       }
@@ -1995,37 +2131,34 @@
     // Boolean - Returns true if a username/ID/UUID/object exists in the userlist.
     // ID - String (username or user object)
     checkForID(args) {
-
       // Legacy ulist handling
       if (clVars.ulist.includes(args.ID)) return true;
 
       // New ulist handling
       if (clVars.linkState.identifiedProtocol > 2) {
         if (this.isValidJSON(args.ID)) {
-          return clVars.ulist.some(o => (
-            (o.username === JSON.parse(args.ID).username)
-            && 
-            (o.id == JSON.parse(args.ID).id)
-          ));
+          return clVars.ulist.some(
+            (o) =>
+              o.username === JSON.parse(args.ID).username &&
+              o.id == JSON.parse(args.ID).id
+          );
         } else {
-          return clVars.ulist.some(o => (
-            (o.username === String(args.ID))
-            || 
-            (o.id == args.ID)
-          ));
+          return clVars.ulist.some(
+            (o) => o.username === String(args.ID) || o.id == args.ID
+          );
         }
       } else return false;
     }
 
     // Boolean - Returns true if the input JSON is valid.
     // JSON_STRING - String
-    isValidJSON(args) { 
+    isValidJSON(args) {
       try {
         JSON.parse(args.JSON_STRING);
         return true;
       } catch {
         return false;
-      };
+      }
     }
 
     // Command - Establishes a connection to a server.
@@ -2034,7 +2167,7 @@
       if (clVars.socket != null) {
         console.warn("[CloudLink] Already connected to a server.");
         return;
-      };
+      }
       return newClient(args.IP);
     }
 
@@ -2044,11 +2177,20 @@
       if (clVars.socket != null) {
         console.warn("[CloudLink] Already connected to a server.");
         return;
-      };
-      if (!clVars.serverList.hasOwnProperty(String(args.ID))) {
+      }
+      // This is the only server that's listed and works.
+      if (Scratch.Cast.toNumber(args.ID) >= 1) {
+        args.ID = 7;
+      }
+      if (
+        !Object.prototype.hasOwnProperty.call(
+          clVars.serverList,
+          String(args.ID)
+        )
+      ) {
         console.warn("[CloudLink] Not a valid server ID!");
         return;
-      };
+      }
       return newClient(clVars.serverList[String(args.ID)]["url"]);
     }
 
@@ -2057,7 +2199,7 @@
       if (clVars.socket == null) {
         console.warn("[CloudLink] Already disconnected.");
         return;
-      };
+      }
       console.log("[CloudLink] Disconnecting...");
       clVars.linkState.isAttemptingGracefulDisconnect = true;
       clVars.socket.close(1000, "Client going away");
@@ -2073,13 +2215,13 @@
       if (clVars.username.attempted) {
         console.warn("[CloudLink] Already attempting to set username!");
         return;
-      };
+      }
 
       // Prevent running if the username is already set.
       if (clVars.username.accepted) {
         console.warn("[CloudLink] Already set username!");
         return;
-      };
+      }
 
       // Update state
       clVars.username.attempted = true;
@@ -2092,28 +2234,31 @@
     // Command - Prepares the next transmitted message to have a listener ID attached to it.
     // ID - String (listener ID)
     createListener(args) {
-
       // Must be connected to set a username.
       if (clVars.socket == null) return;
-      
+
       // Require server support
       if (clVars.linkState.identifiedProtocol < 2) {
-        console.warn("[CloudLink] Server is too old! Must be at least 0.1.8.x to support listeners.");
+        console.warn(
+          "[CloudLink] Server is too old! Must be at least 0.1.8.x to support listeners."
+        );
         return;
       }
 
       // Prevent running if the username hasn't been set.
       if (!clVars.username.accepted) {
-        console.warn("[CloudLink] Username must be set before creating a listener!");
+        console.warn(
+          "[CloudLink] Username must be set before creating a listener!"
+        );
         return;
-      };
+      }
 
       // Must be used once per packet
       if (clVars.listeners.enablerState) {
         console.warn("[CloudLink] Cannot create multiple listeners at a time!");
         return;
       }
-      
+
       // Update state
       clVars.listeners.enablerState = true;
       clVars.listeners.enablerValue = args.ID;
@@ -2121,34 +2266,37 @@
 
     // Command - Subscribes to various rooms on a server.
     // ROOMS - String (JSON Array or single string)
-    linkToRooms(args) { 
-
+    linkToRooms(args) {
       // Must be connected to set a username.
       if (clVars.socket == null) return;
 
       // Require server support
       if (clVars.linkState.identifiedProtocol < 2) {
-        console.warn("[CloudLink] Server is too old! Must be at least 0.1.8.x to support rooms.");
+        console.warn(
+          "[CloudLink] Server is too old! Must be at least 0.1.8.x to support rooms."
+        );
         return;
       }
 
       // Prevent running if the username hasn't been set.
       if (!clVars.username.accepted) {
-        console.warn("[CloudLink] Username must be set before linking to rooms!");
+        console.warn(
+          "[CloudLink] Username must be set before linking to rooms!"
+        );
         return;
-      };
+      }
 
       // Prevent running if already linked.
       if (clVars.rooms.isLinked) {
         console.warn("[CloudLink] Already linked to rooms!");
         return;
-      };
+      }
 
       // Prevent running if a room link is in progress.
       if (clVars.rooms.isAttemptingLink) {
         console.warn("[CloudLink] Currently linking to rooms! Please wait!");
         return;
-      };
+      }
 
       clVars.rooms.isAttemptingLink = true;
       sendMessage({ cmd: "link", val: args.ROOMS, listener: "link" });
@@ -2156,68 +2304,80 @@
 
     // Command - Specifies specific subscribed rooms to transmit messages to.
     // ROOMS - String (JSON Array or single string)
-    selectRoomsInNextPacket(args) { 
-
+    selectRoomsInNextPacket(args) {
       // Must be connected to user rooms.
       if (clVars.socket == null) return;
 
       // Require server support
       if (clVars.linkState.identifiedProtocol < 2) {
-        console.warn("[CloudLink] Server is too old! Must be at least 0.1.8.x to support rooms.");
+        console.warn(
+          "[CloudLink] Server is too old! Must be at least 0.1.8.x to support rooms."
+        );
         return;
       }
 
       // Prevent running if the username hasn't been set.
       if (!clVars.username.accepted) {
-        console.warn("[CloudLink] Username must be set before selecting rooms!");
+        console.warn(
+          "[CloudLink] Username must be set before selecting rooms!"
+        );
         return;
-      };
+      }
 
       // Require once per packet
       if (clVars.rooms.enablerState) {
-        console.warn("[CloudLink] Cannot use the room selector more than once at a time!");
+        console.warn(
+          "[CloudLink] Cannot use the room selector more than once at a time!"
+        );
         return;
       }
 
       // Prevent running if not linked.
       if (!clVars.rooms.isLinked) {
-        console.warn("[CloudLink] Cannot use room selector while not linked to rooms!");
+        console.warn(
+          "[CloudLink] Cannot use room selector while not linked to rooms!"
+        );
         return;
-      };
+      }
 
       clVars.rooms.enablerState = true;
       clVars.rooms.enablerValue = args.ROOMS;
-    } 
+    }
 
     // Command - Unsubscribes from all rooms and re-subscribes to the the "default" room on the server.
-    unlinkFromRooms() { 
-
+    unlinkFromRooms() {
       // Must be connected to user rooms.
       if (clVars.socket == null) return;
 
       // Require server support
       if (clVars.linkState.identifiedProtocol < 2) {
-        console.warn("[CloudLink] Server is too old! Must be at least 0.1.8.x to support rooms.");
+        console.warn(
+          "[CloudLink] Server is too old! Must be at least 0.1.8.x to support rooms."
+        );
         return;
       }
 
       // Prevent running if the username hasn't been set.
       if (!clVars.username.accepted) {
-        console.warn("[CloudLink] Username must be set before unjoining rooms!");
+        console.warn(
+          "[CloudLink] Username must be set before unjoining rooms!"
+        );
         return;
-      };
+      }
 
       // Prevent running if already unlinked.
       if (!clVars.rooms.isLinked) {
         console.warn("[CloudLink] Already unlinked from rooms!");
         return;
-      };
+      }
 
       // Prevent running if a room unlink is in progress.
       if (clVars.rooms.isAttemptingUnlink) {
-        console.warn("[CloudLink] Currently unlinking from rooms! Please wait!");
+        console.warn(
+          "[CloudLink] Currently unlinking from rooms! Please wait!"
+        );
         return;
-      };
+      }
 
       clVars.rooms.isAttemptingUnlink = true;
       sendMessage({ cmd: "unlink", val: "", listener: "unlink" });
@@ -2226,7 +2386,6 @@
     // Command - Sends a gmsg value.
     // DATA - String
     sendGData(args) {
-
       // Must be connected.
       if (clVars.socket == null) return;
 
@@ -2236,15 +2395,16 @@
     // Command - Sends a pmsg value.
     // DATA - String, ID - String (recipient ID)
     sendPData(args) {
-
       // Must be connected.
       if (clVars.socket == null) return;
 
       // Prevent running if the username hasn't been set.
       if (!clVars.username.accepted) {
-        console.warn("[CloudLink] Username must be set before sending private messages!");
+        console.warn(
+          "[CloudLink] Username must be set before sending private messages!"
+        );
         return;
-      };
+      }
 
       sendMessage({ cmd: "pmsg", val: args.DATA, id: args.ID });
     }
@@ -2252,7 +2412,6 @@
     // Command - Sends a gvar value.
     // DATA - String, VAR - String (variable name)
     sendGDataAsVar(args) {
-
       // Must be connected.
       if (clVars.socket == null) return;
 
@@ -2262,15 +2421,16 @@
     // Command - Sends a pvar value.
     // DATA - String, VAR - String (variable name), ID - String (recipient ID)
     sendPDataAsVar(args) {
-
       // Must be connected.
       if (clVars.socket == null) return;
 
       // Prevent running if the username hasn't been set.
       if (!clVars.username.accepted) {
-        console.warn("[CloudLink] Username must be set before sending private variables!");
+        console.warn(
+          "[CloudLink] Username must be set before sending private variables!"
+        );
         return;
-      };
+      }
 
       sendMessage({ cmd: "pvar", val: args.DATA, name: args.VAR, id: args.ID });
     }
@@ -2278,7 +2438,6 @@
     // Command - Sends a raw-format command without specifying an ID.
     // CMD - String (command), DATA - String
     runCMDnoID(args) {
-
       // Must be connected.
       if (clVars.socket == null) return;
 
@@ -2288,15 +2447,16 @@
     // Command - Sends a raw-format command with an ID.
     // CMD - String (command), DATA - String, ID - String (recipient ID)
     runCMD(args) {
-
       // Must be connected.
       if (clVars.socket == null) return;
 
       // Prevent running if the username hasn't been set.
       if (!clVars.username.accepted) {
-        console.warn("[CloudLink] Username must be set before using this command!");
+        console.warn(
+          "[CloudLink] Username must be set before using this command!"
+        );
         return;
-      };
+      }
 
       sendMessage({ cmd: args.CMD, val: args.DATA, id: args.ID });
     }
@@ -2305,16 +2465,16 @@
     // TYPE - String (menu datamenu)
     resetNewData(args) {
       switch (args.TYPE) {
-        case 'Global data':
+        case "Global data":
           clVars.gmsg.hasNew = false;
           break;
-        case 'Private data':
+        case "Private data":
           clVars.pmsg.hasNew = false;
           break;
-        case 'Direct data':
+        case "Direct data":
           clVars.direct.hasNew = false;
           break;
-        case 'Status code':
+        case "Status code":
           clVars.statuscode.hasNew = false;
           break;
       }
@@ -2324,25 +2484,46 @@
     // TYPE - String (menu varmenu), VAR - String (variable name)
     resetNewVarData(args) {
       switch (args.TYPE) {
-        case 'Global variables':
-          if (!clVars.gvar.varStates.hasOwnProperty(String(args.VAR))) {
-            console.warn(`[CloudLink] Global variable ${args.VAR} does not exist!`);
+        case "Global variables":
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.gvar.varStates,
+              String(args.VAR)
+            )
+          ) {
+            console.warn(
+              `[CloudLink] Global variable ${args.VAR} does not exist!`
+            );
             return;
           }
           clVars.gvar.varStates[String(args.ID)].hasNew = false;
-        case 'Private variables':
-          if (!clVars.pvar.varStates.hasOwnProperty(String(args.VAR))) {
-            console.warn(`[CloudLink] Private variable ${args.VAR} does not exist!`);
+          break;
+        case "Private variables":
+          if (
+            !Object.prototype.hasOwnProperty.call(
+              clVars.pvar.varStates,
+              String(args.VAR)
+            )
+          ) {
+            console.warn(
+              `[CloudLink] Private variable ${args.VAR} does not exist!`
+            );
             return false;
           }
           clVars.pvar.varStates[String(args.ID)].hasNew = false;
+          break;
       }
     }
 
     // Command - Resets the "returnIsNewListener" boolean state.
     // ID - Listener ID
     resetNewListener(args) {
-      if (!clVars.listeners.varStates.hasOwnProperty(String(args.ID))) {
+      if (
+        !Object.prototype.hasOwnProperty.call(
+          clVars.listeners.varStates,
+          String(args.ID)
+        )
+      ) {
         console.warn(`[CloudLink] Listener ID ${args.ID} does not exist!`);
         return;
       }
@@ -2353,25 +2534,25 @@
     // TYPE - String (menu allmenu)
     clearAllPackets(args) {
       switch (args.TYPE) {
-        case 'Global data':
+        case "Global data":
           clVars.gmsg.queue = [];
           break;
-        case 'Private data':
+        case "Private data":
           clVars.pmsg.queue = [];
           break;
-        case 'Direct data':
+        case "Direct data":
           clVars.direct.queue = [];
           break;
-        case 'Status code':
+        case "Status code":
           clVars.statuscode.queue = [];
           break;
-        case 'Global variables':
+        case "Global variables":
           clVars.gvar.queue = [];
           break;
-        case 'Private variables':
+        case "Private variables":
           clVars.pvar.queue = [];
           break;
-        case 'All data':
+        case "All data":
           clVars.gmsg.queue = [];
           clVars.pmsg.queue = [];
           clVars.direct.queue = [];
